requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
Django==5.1.3
django-auditlog==3.0.0
django-axes==7.0.0
django-cachalot==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-multitenant==4.1.1
django-redis==5.4.0
django-threadlocals==0.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-flex-fields==1.0.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pika==1.3.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
sentry-sdk==2.18.0
serpy==0.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3